RefuseClassification/requirements.txt
matplotlib==3.7.2
numpy==2.0.0
opencv_python==4.9.0.80
Pillow==9.5.0
PySimpleGUI==4.60.5
scikit_learn==1.3.0
seaborn==0.13.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
